Refactored name to match standards
